# static/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: A basic e-commerce api to practice backend development skills
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
  /users:
    get:
      summary: Returns all users.
      description: Retrieve a list of all users in the system.
      responses:
        '200':
          description: A JSON array of users.
          contennt:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Creates a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserRequest'
              required:
                - username
                - email
                - password
                - display_name
      responses:
        '200':
          description: User created successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{username}:
    parameters:
      - in: path
        name: username
        schema:
          type: string
        required: true
        description: Username of the user.
    get:
      summary: Return the user with the given username.
      description: Retrieve the user with the given username in the system.
      responses:
        '200':
          description: A JSON object of the user.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
    delete:
      summary: Delete and return the user with the given username.
      description: Delete the user with the given username in the system.
      responses:
        '200':
          description: User deleted successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
    put:
      summary: Updates the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserRequest'
              required:
                - username
                - email
                - password
                - display_name
                - is_active
      responses:
        '200':
          description: User updated successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
    patch:
      summary: Update (patch) the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
  /stores:
    get:
      summary: Returns all stores.
      description: Retrieve a list of all stores in the system.
      responses:
        '200':
          description: A JSON array of stores.
          contennt:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Store'
    post:
      summary: Creates a new store.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/StoreRequest'
              required:
                - Name
                - Slug 
      responses:
        '200':
          description: Store created successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
  /stores/{slug}:
    parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
        description: Slug of the store.
    get:
      summary: Return the store with the given slug.
      description: Retrieve the store with the given slug in the system.
      responses:
        '200':
          description: A JSON object of the store.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '404':
          description: No store with the given slug found.
    delete:
      summary: Delete and return the store with the given slug.
      description: Delete the store with the given slug in the system.
      responses:
        '200':
          description: Store deleted successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '404':
          description: No store with the given slug found.
    put:
      summary: Updates the store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/StoreRequest'
              required:
                - Name
                - Slug 
                - Description
      responses:
        '200':
          description: User updated successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
    patch:
      summary: Update (patch) the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRequest'
      responses:
        '200':
          description: User updated successfully.
          contennt:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No user with the given username found.
  /users/stores:
    get:
      summary: Returns users and their stores
      description:  Retrieve a list of users, each with their associated stores.
      responses:
        '200':
          description: A list of users and their stores.
          contennt:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserStores'

components:
  schemas:
    User:
      type: object
      properties:
        Username:
          type: string
          example: neo
        Email:
          type: string
          format: email
          example: tanderson@metacortex.com
        DisplayName:
          type: string
          example: Thomas A. Anderson
        IsActive:
          type: boolean
          example: true
        CreatedAt:
          type: string
          format: date-time
          example: "2026-02-25T10:00:03Z"
        UpdatedAt:
          type: string
          format: date-time
          example: "2026-02-25T12:00:03Z"
      required:
        - Username
        - Email
        - DisplayName
        - IsActive
        - CreatedAt
        - UpdatedAt
    UserRequest:
      type: object
      properties:
        username:
          type: string
          example: neo
        email:
          type: string
          format: email
          example: tanderson@metacortex.com
        password:
          type: string
          example: very$ecretpassword123
        display_name:
          type: string
          example: Thomas A. Anderson
        is_active:
          type: boolean
          example: true
    Store:
      type: object
      properties:
        Name:
          type: string
          example: MetaCortex
        Slug:
          type: string
          example: meta-cortex
        Description:
          type: string
          example: "MetaCortex is a leading software solutions provider specializing in enterprise applications, developer tools, and IT services, delivering innovative technology for modern businesses."
        CreatedAt:
          type: string
          format: date-time
          example: "2026-02-25T10:00:03Z"
      required:
        - Name
        - Slug       
        - Description
        - CreatedAt
    StoreRequest:
      type: object
      properties:
        Name:
          type: string
          example: MetaCortex
        Slug:
          type: string
          example: meta-cortex
        Description:
          type: string
          example: "MetaCortex is a leading software solutions provider specializing in enterprise applications, developer tools, and IT services, delivering innovative technology for modern businesses."
    UserStores:
      type: object
      properties:
        User:
          $ref: '#/components/schemas/User'
        Stores:
          type: array
          items:
            $ref: '#/components/schemas/Store'
      required:
      - User
      - Stores